+++ core/data/src/main/java/com/depromeet/team5/core/data/datasource/RemoteDataSource.kt
@@ -3,10 +3,8 @@ package com.depromeet.team5.core.data.datasource
 import com.depromeet.team5.core.data.model.FeedbackData
 import com.depromeet.team5.core.data.model.MyPrincipleGroupData
 import com.depromeet.team5.core.data.model.MyPrincipleGroupsInfoData
-import com.depromeet.team5.core.data.model.RetrospectionData
 import com.depromeet.team5.core.data.model.RetrospectionListData
 import com.depromeet.team5.core.data.model.SearchData
-import com.depromeet.team5.core.data.request.CreateRetrospectionRequestData
 import com.depromeet.team5.core.data.model.SystemPrincipleData
 import com.depromeet.team5.core.data.model.UserStatsData
 
@@ -14,8 +12,6 @@ import com.depromeet.team5.core.data.model.UserStatsData
 interface RemoteDataSource {
 
     suspend fun search(query: String): SearchData
-
-    suspend fun createRetrospection(request: CreateRetrospectionRequestData): RetrospectionData
 
     suspend fun createFeedback(retrospectionId: Int): FeedbackData
 
@@ -28,12 +24,6 @@ interface RemoteDataSource {
     suspend fun deletePrinciple(principleId: Int)
 
     suspend fun createPrincipleGroup(body: Map<String, Any?>): MyPrincipleGroupData
-
-    suspend fun uploadImageUri(
-        domain: String,
-        uri: String,
-        fileName: String? = null
-    ): Int
 
     suspend fun userStats(): UserStatsData
