+++ core/data/src/main/java/com/depromeet/team5/core/data/repositoryimpl/HedgeRepositoryImpl.kt
@@ -1,16 +1,13 @@
 package com.depromeet.team5.core.data.repositoryimpl
 
 import com.depromeet.team5.core.data.datasource.RemoteDataSource
-import com.depromeet.team5.core.data.mapper.toData
 import com.depromeet.team5.core.domain.model.Feedback
 import com.depromeet.team5.core.domain.model.MyPrincipleGroup
-import com.depromeet.team5.core.domain.model.Retrospection
 import com.depromeet.team5.core.domain.model.RetrospectionList
 import com.depromeet.team5.core.domain.model.Search
 import com.depromeet.team5.core.domain.model.SystemPrinciple
 import com.depromeet.team5.core.domain.model.UserStats
 import com.depromeet.team5.core.domain.repository.HedgeRepository
-import com.depromeet.team5.core.domain.request.CreateRetrospectionRequest
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.map
@@ -23,10 +20,6 @@ internal class HedgeRepositoryImpl @Inject constructor(
 
     override fun search(query: String): Flow<Search> = flow {
         emit(remoteDataSource.search(query).toDomain())
-    }
-
-    override fun createRetrospection(request: CreateRetrospectionRequest): Flow<Retrospection> = flow {
-        emit(remoteDataSource.createRetrospection(request.toData()).toDomain())
     }
 
     override fun createFeedback(
@@ -56,12 +49,6 @@ internal class HedgeRepositoryImpl @Inject constructor(
         emit(remoteDataSource.createPrincipleGroup(body))
     }
         .map { it.toDomain() }
-
-    override suspend fun uploadImageUri(
-        domain: String,
-        uri: String,
-        fileName: String?
-    ): Int = remoteDataSource.uploadImageUri(domain, uri, fileName)
 
     override fun userStats(): Flow<UserStats> = flow {
         emit(remoteDataSource.userStats().toDomain())
