+++ core/domain/src/main/java/com/depromeet/team5/core/domain/repository/RetrospectionRepository.kt
@@ -3,14 +3,23 @@ package com.depromeet.team5.core.domain.repository
 import com.depromeet.team5.core.domain.model.BaseDomain
 import com.depromeet.team5.core.domain.model.Memo
 import com.depromeet.team5.core.domain.model.Retrospection
+import com.depromeet.team5.core.domain.request.CreateRetrospectionRequest
 import kotlinx.coroutines.flow.Flow
 
 
 interface RetrospectionRepository {
 
+    fun createRetrospection(request: CreateRetrospectionRequest): Flow<BaseDomain<Retrospection>>
+
     fun getRetrospection(retrospectionId: Int): Flow<BaseDomain<Retrospection>>
 
     fun deleteRetrospection(retrospectionId: Int): Flow<BaseDomain<String>>
+
+    suspend fun uploadImageUri(
+        domain: String,
+        uri: String,
+        fileName: String? = null
+    ): Int
 
     fun createMemo(
         retrospectionId: Int,
