+++ core/remotedatasource/src/main/java/com/depromeet/team5/core/remotedatasource/RemoteDataSourceImpl.kt
@@ -5,14 +5,11 @@ import com.depromeet.team5.core.data.datasource.RemoteDataSource
 import com.depromeet.team5.core.data.model.FeedbackData
 import com.depromeet.team5.core.data.model.MyPrincipleGroupData
 import com.depromeet.team5.core.data.model.MyPrincipleGroupsInfoData
-import com.depromeet.team5.core.data.model.RetrospectionData
 import com.depromeet.team5.core.data.model.RetrospectionListData
 import com.depromeet.team5.core.data.model.SearchData
-import com.depromeet.team5.core.data.request.CreateRetrospectionRequestData
 import com.depromeet.team5.core.data.model.SystemPrincipleData
 import com.depromeet.team5.core.data.model.UserStatsData
 import com.depromeet.team5.core.remotedatasource.apisource.HedgeApiSource
-import com.depromeet.team5.core.remotedatasource.mapper.toRemoteData
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -24,19 +21,6 @@ internal class RemoteDataSourceImpl @Inject constructor(
 
     override suspend fun search(query: String): SearchData =
         hedgeApiSource.search(query).toData()
-
-    override suspend fun createRetrospection(request: CreateRetrospectionRequestData): RetrospectionData =
-        hedgeApiSource.createRetrospection(request.toRemoteData()).toData()
-
-    override suspend fun uploadImageUri(
-        domain: String,
-        uri: String,
-        fileName: String?
-    ) = hedgeApiSource.uploadImageUri(
-        domain = domain,
-        uri = uri.toUri(),
-        fileName = fileName
-    )
 
     override suspend fun createFeedback(retrospectionId: Int): FeedbackData =
         hedgeApiSource.createFeedback(retrospectionId).toData()
