+++ core/remotedatasource/src/main/java/com/depromeet/team5/core/remotedatasource/apisource/HedgeApiSource.kt
@@ -1,22 +1,17 @@
 package com.depromeet.team5.core.remotedatasource.apisource
 
-import android.net.Uri
 import com.depromeet.team5.core.remotedatasource.model.FeedbackRemoteData
 import com.depromeet.team5.core.remotedatasource.model.MyPrincipleGroupRemoteData
 import com.depromeet.team5.core.remotedatasource.model.MyPrincipleGroupsInfoRemoteData
 import com.depromeet.team5.core.remotedatasource.model.RetrospectionListRemoteData
-import com.depromeet.team5.core.remotedatasource.model.RetrospectionRemoteData
 import com.depromeet.team5.core.remotedatasource.model.SearchRemoteData
 import com.depromeet.team5.core.remotedatasource.model.SystemPrincipleRemoteData
 import com.depromeet.team5.core.remotedatasource.model.UserStatsRemoteData
-import com.depromeet.team5.core.remotedatasource.request.CreateRetrospectionRequestRemoteData
 
 
 interface HedgeApiSource {
 
     suspend fun search(query: String): SearchRemoteData
-
-    suspend fun createRetrospection(request: CreateRetrospectionRequestRemoteData): RetrospectionRemoteData
 
     suspend fun createFeedback(retrospectionId: Int): FeedbackRemoteData
 
@@ -29,12 +24,6 @@ interface HedgeApiSource {
     suspend fun deletePrinciple(principleId: Int)
 
     suspend fun createPrincipleGroup(body: Map<String, Any?>): MyPrincipleGroupRemoteData
-
-    suspend fun uploadImageUri(
-        domain: String,
-        uri: Uri,
-        fileName: String? = null
-    ): Int
 
     suspend fun userStats(): UserStatsRemoteData
