+++ core/retrofit/src/main/java/com/depromeet/team5/core/retrofit/api/RetrospectionApi.kt
@@ -1,16 +1,24 @@
 package com.depromeet.team5.core.retrofit.api
 
+import com.depromeet.team5.core.remotedatasource.request.CreateRetrospectionRequestRemoteData
 import com.depromeet.team5.core.retrofit.model.BaseResponse
+import com.depromeet.team5.core.retrofit.model.ImageUploadResponse
 import com.depromeet.team5.core.retrofit.model.MemoResponse
 import com.depromeet.team5.core.retrofit.model.RetrospectionResponse
+import okhttp3.MultipartBody
 import retrofit2.http.Body
 import retrofit2.http.DELETE
 import retrofit2.http.GET
+import retrofit2.http.Multipart
 import retrofit2.http.POST
 import retrofit2.http.PUT
+import retrofit2.http.Part
 import retrofit2.http.Path
 
 interface RetrospectionApi {
+
+    @POST(PATH)
+    suspend fun createRetrospection(@Body body: CreateRetrospectionRequestRemoteData): BaseResponse<RetrospectionResponse>
 
     @GET("$PATH/{retrospectionId}")
     suspend fun getRetrospection(
@@ -21,6 +29,13 @@ interface RetrospectionApi {
     suspend fun deleteRetrospection(
         @Path("retrospectionId") retrospectionId: Int,
     ): BaseResponse<String>
+
+    @Multipart
+    @POST("api/v1/{domain}/images/upload")
+    suspend fun imageUpload(
+        @Path("domain") domain: String,
+        @Part file: MultipartBody.Part,
+    ): BaseResponse<ImageUploadResponse>
 
     @POST("$PATH/{retrospectionId}/memos")
     suspend fun createMemo(
