+++ features/feedback/src/main/java/com/depromeet/team5/features/feedback/screen/AiFeedbackViewModel.kt
@@ -40,7 +40,7 @@ class AiFeedbackViewModel @Inject constructor(
                 request = request.copy(orderDate = formatDate(request.orderDate)),
                 principles = principleGroupState.principles,
             )
-                .flatMapConcat { createFeedbackUseCase(it.id) }
+                .flatMapConcat { createFeedbackUseCase(it.data!!.id) }
                 .map {
                     if (it.data != null) {
                         AiFeedbackUiState.Success(
